Feat/profile: profile 추가

+++ backend/src/main/java/com/healthsync/project/calc/dto/MetricsResponse.java
@@ -2,6 +2,8 @@ package com.healthsync.project.calc.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -9,11 +11,11 @@ import lombok.*;
 @Builder
 public class MetricsResponse {
 
-    private double bmi;               // 계산된 BMI 값
-    private String category;          // BMI 분류
-    private double standardWeight;    // 표준체중
-    private double dailyCalories;     // 하루 권장 칼로리
-    private double bmr;               // 기초대사량
-    private double activityCalories;  // 활동대사량
+    private BigDecimal bmi;               // 계산된 BMI 값
+    private String category;              // BMI 분류
+    private BigDecimal standardWeight;    // 표준체중
+    private BigDecimal dailyCalories;     // 하루 권장 칼로리
+    private BigDecimal bmr;               // 기초대사량
+    private BigDecimal activityCalories;  // 활동대사량
 
 }
